Initialize CameraPreview.mPreviewing with false
We don't explicitly setup preview in constructor. Although in samples
startCamera is called shortly after CameraPreview creation.

The reason showCameraPreview gets called in samples is that
surfaceChanged explicitly stops and shows the preview, which
makes mPreviewing to be consistent. And it so happens that
surfaceChanged happens in samples.

But in scenario when startCamera/initCameraPreview is called
slightly after activity/fragment onCreate there are no more
surfaceChanged events coming and initCameraPreview thinks
that preview is up and running and doesn't actually start it.
Hence a black screen/no preview.